+++ score_analysis/__init__.py
@@ -1,5 +1,5 @@
 from . import metrics, tools
-from .cm import BinaryConfusionMatrix, ConfusionMatrix
+from .cm import ConfusionMatrix
 from .scores import Scores
 
 __version__ = "0.0.5"